Make things still work when the mchain and id chain are in a different order
This one's odd. We allow gaps in the chain, and we don't know
something's been deleted till the block in fron "overshoots" it with
it's back pointer in the ID chain.

This is fine because we eventually get the full chain because the mchain
stays a true chain on any given node.